mcp/requirements.txt
mcp
python-dotenv
pytest
pubchempy
mp-api
pymatgen
ase
matminer
lxml  # optional: used for pretty-printing fixed URDF XML
# enumlib (enum.x) is required for pymatgen_enumeration_generator but is NOT on PyPI.
# It is installed automatically by setup.sh (via conda) on Linux/macOS/WSL.
# Manual install: conda install -c conda-forge enumlib
# Windows without WSL: not supported — see README.md.